requirements.txt
gradio>=4.0.0
transformers
pandas
matplotlib
faiss-cpu
python-dotenv
# Optional: chromadb, promptfoo, openllmetry, helicone
